requirements.txt
# LLM Provider
langchain-groq==0.2.1

# Core LangChain
langchain==0.3.11
langchain-community==0.3.10
torch==2.4.1

# Vector Store & Embeddings
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1

# Utilities
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.2.1
pydantic==2.10.3
python-dateutil==2.9.0

# Optional: UI
streamlit==1.41.1

# Development
pytest==8.3.4